+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `CorticalParcellationStats.read` support `pathlib.Path`, `"http://…"`, `"https://…"`, `"s3://…"` etc.
+  via `pandas.io.common.get_filepath_or_buffer`
+  (https://github.com/fphammerle/freesurfer-stats/issues/6)
 
 ## [1.1.1] - 2020-05-07
 ### Fixed